requirements.txt
feedparser==6.0.10
requests==2.31.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
schedule==1.2.1
